+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 ## Change Log
 
+### 1.2.0 (2018/12/05)
+- Select episodes by hyphen-sequence
+
 ### 1.1.1 (2018/12/04)
 - Fixed VTT+CSS parsing with some subtitles
 
 ### 1.1.0 (2018/12/01)
 - `stag` option for subtitles, updated configuration file
-- removed `dateformat` module
+- Removed `dateformat` module
 
 ### 1.0.0 (2018/11/30)
 - First public release